deepen: delete dead services/filing/ (singular) orbital
Replaces the 9-file services/filing/ directory with the small file-level
surface production actually exercises under services/filings/ (plural).

The singular directory paralleled the live plural one — every functional
file had a live equivalent under services/filings/ (databaseService →
database/filingDatabase, fallbackSummary → summaries/fallbackSummaryGenerator,
formTypeService → utils/formTypeUtils, sendEmailSummary → email/sendEmailSummary,
summaryGenerationService → summaries/filingSummaryService). Only
services/filing/types.ts had production callers (7 files under services/filings/
importing FilingSummaryResult / FilingError / SECFiling / Company) — a
zero-caller orbital cluster kept alive by one shared types module.

Deletion collapses a hypothetical seam (8 candidate adapters under
services/filing/, all with 0 non-test production callers) to the small
file-level surface production actually exercises. Types migrate to a new
services/filings/shared-types.ts holding only the 4 types with real
callers; the parallel incomplete services/filings/summaries/types.ts
(FilingSummaryResult missing databaseId/isCacheHit/cacheUsageCount/
qualityScore, sole importer services/filingService.ts) is deleted in the
same move so one canonical shape covers every services/filings/ consumer.

Tests that only exercised the dead orbital were deleted (unit,
integration, performance, security, edge-case suites of the
summaryGenerationService and sendEmailSummary variants). Two tests that
merely mocked the dead paths have their mock blocks removed; behaviour
covered elsewhere. tools/test-xml-monitoring.{ts,js,sh} referenced
services/filing/getFilingById and are removed with it.

Follows the pattern of #849, #834, #808, #809, #777 — delete dead
orbital rather than deepen an already-live seam.

Diff: 34 files changed, ~10 insertions, ~7,495 deletions.
