--- services/filing/index.ts
@@ -1,5 +1,0 @@
-// Export all filing service functionality from this index file
-export * from './types';
-export * from './getFilingById';
-export * from './sendEmailSummary';
-export * from './utils';
