--- services/filing/utils.ts
@@ -1,371 +1,0 @@
-import { FilingSummary } from '../../types/sec/filing';
-import { FilingSummaryResult, FilingError } from './types';
-import { FormTypeMetadata, getFormMetadata } from '../../lib/sec-edgar/form-registry';
-import { ParsedContent } from '../../lib/parsers/form-parser';
-import { FilingType } from '../../types/sec/filing';
-
-/**
- * Sanitize text for email content by removing problematic characters
- * @param text Text to sanitize
- * @returns Sanitized text
- */
-export function sanitizeForEmail(text: string): string {
-  if (!text) return '';
-  
-  // Replace problematic characters that might cause issues with email rendering
-  return text
-    .replace(/[\u0000-\u001F\u007F-\u009F]/g, '') // Remove control characters
-    .replace(/[\u2028\u2029]/g, ' ') // Replace line/paragraph separators
-    .replace(/[\uFFFE\uFFFF\uFFF0-\uFFFD]/g, '') // Remove Unicode BOM and specials
-    .replace(/[\u0080-\u00A0]/g, ' ') // Replace Latin-1 control chars with spaces
-    .trim();
-}
-
-/**
- * Generate a plain text version of the email
- * @param summaries List of filing summaries
- * @param errors List of errors encountered
- * @returns Plain text email content
- */
-export function generatePlainTextEmail(
-  summaries: FilingSummary[] | FilingSummaryResult[], 
-  errors: FilingError[] = []
-): string {
-  let text = `SEC Filing Summaries - ${new Date().toLocaleDateString()}\n\n`;
-  
-  // Group by ticker
-  const tickerMap: Record<string, FilingSummary[] | FilingSummaryResult[]> = {};
-  
-  summaries.forEach(summary => {
-    if (!tickerMap[summary.ticker]) {
-      tickerMap[summary.ticker] = [];
-    }
-    tickerMap[summary.ticker].push(summary);
-  });
-  
-  // Generate text for each ticker
-  Object.entries(tickerMap).forEach(([ticker, tickerSummaries]) => {
-    text += `${ticker} - ${tickerSummaries[0].companyName}\n`;
-    text += '='.repeat(40) + '\n\n';
-    
-    tickerSummaries.forEach(summary => {
-      const formMetadata = getFormMetadata(summary.filingType);
-      const formName = formMetadata ? formMetadata.displayName : summary.filingType;
-      const filingDate = new Date(summary.filingDate).toLocaleDateString();
-      
-      text += `${formName} - Filed on ${filingDate}\n\n`;
-      text += `${summary.summaryText}\n\n`;
-      
-      if (summary.keyPoints && summary.keyPoints.length > 0) {
-        text += 'Key Points:\n';
-        summary.keyPoints.forEach(point => {
-          text += `- ${point}\n`;
-        });
-        text += '\n';
-      }
-      
-      text += `View on SEC Website: ${summary.url}\n\n`;
-    });
-  });
-  
-  // Add errors if any
-  if (errors.length > 0) {
-    text += '\nIssues Encountered:\n';
-    errors.forEach(err => {
-      text += `${err.ticker}: ${err.error}\n`;
-    });
-  }
-  
-  // Add footer
-  text += '\n--\n';
-  text += 'This email was generated by tldrSEC. The information provided is for informational purposes only and should not be considered financial advice.\n';
-  text += `© ${new Date().getFullYear()} tldrSEC`;
-  
-  return text;
-}
-
-/**
- * Generate HTML email content for filing summaries
- * @param summaries List of filing summaries
- * @param errors List of errors encountered
- * @returns HTML email content
- */
-export function generateEmailHtml(
-  summaries: FilingSummaryResult[], 
-  errors: FilingError[] = []
-): string {
-  let html = `
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-      <meta charset="UTF-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <title>SEC Filing Summaries</title>
-      <style>
-        /* Base styles */
-        body { 
-          font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif; 
-          line-height: 1.6; 
-          color: #333; 
-          max-width: 800px; 
-          margin: 0 auto;
-          padding: 20px;
-          background-color: #f9fafb;
-        }
-        
-        /* Header */
-        .header { 
-          background: linear-gradient(135deg, #7C3AED 0%, #EC4899 100%);
-          color: white; 
-          padding: 30px 20px; 
-          text-align: center;
-          border-radius: 12px 12px 0 0;
-          margin-bottom: 30px;
-        }
-        
-        .header h1 { margin: 0 0 10px; font-size: 28px; }
-        
-        /* Summary cards */
-        .summary { 
-          background-color: white;
-          border-radius: 12px; 
-          box-shadow: 0 4px 6px rgba(0,0,0,0.05);
-          padding: 25px; 
-          margin-bottom: 25px;
-        }
-        
-        .summary h2 { 
-          margin-top: 0; 
-          color: #4F46E5;
-          font-size: 20px;
-          border-bottom: 1px solid #E5E7EB;
-          padding-bottom: 10px;
-        }
-        
-        .meta { 
-          color: #6B7280; 
-          font-size: 14px; 
-          margin-bottom: 15px;
-        }
-        
-        /* Key points section */
-        .key-points { 
-          background-color: #F3F4F6; 
-          padding: 15px; 
-          border-left: 4px solid #4F46E5;
-          border-radius: 0 8px 8px 0;
-          margin: 20px 0;
-        }
-        
-        .key-points h3 { 
-          margin-top: 0; 
-          color: #4B5563;
-          font-size: 16px;
-        }
-        
-        .key-points ul { margin: 10px 0 0; padding-left: 20px; }
-
-        .summary-text { margin-bottom: 15px; font-size: 14px; color: #374151; }
-
-        /* Button styling */
-        .filing-link { 
-          display: inline-block;
-          margin-top: 15px;
-          padding: 12px 20px;
-          background: linear-gradient(135deg, #7C3AED 0%, #EC4899 100%);
-          color: white;
-          text-decoration: none;
-          border-radius: 8px;
-          font-size: 14px;
-          font-weight: bold;
-        }
-        .filing-link:hover { opacity: 0.9; }
-
-        /* Errors styling */
-        .errors { background-color: #fff0f0; padding: 15px; margin-top: 20px; border-radius: 8px; }
-
-        /* Footer */
-        .footer { margin-top: 30px; text-align: center; font-size: 12px; color: #6B7280; }
-      </style>
-    </head>
-    <body>
-      <div class="header">
-        <h1>SEC Filing Summaries</h1>
-        <p>${new Date().toLocaleDateString()}</p>
-      </div>
-  `;
-  
-  // Add summaries
-  summaries.forEach(summary => {
-    const formMetadata = getFormMetadata(summary.filingType);
-    const formName = formMetadata ? formMetadata.displayName : summary.filingType;
-    const filingDate = new Date(summary.filingDate).toLocaleDateString();
-    
-    // Make sure we have a summary text
-    const summaryTextContent = summary.summaryText && summary.summaryText.trim() !== '' ? 
-      sanitizeForEmail(summary.summaryText) : 
-      `This is a ${formName} filing from ${summary.companyName}. View the original filing for complete details.`;
-    
-    html += `
-      <div class="summary">
-        <h2>${sanitizeForEmail(summary.companyName)} (${sanitizeForEmail(summary.ticker)}) - ${sanitizeForEmail(formName)}</h2>
-        <div class="meta">Filed on: ${filingDate}</div>
-        
-        <div class="summary-text">
-          <p>${summaryTextContent}</p>
-        </div>
-        
-        <div class="key-points">
-          <h3>Key Points</h3>
-          <ul>
-            ${summary.keyPoints.map((point: string) => `<li>${sanitizeForEmail(point)}</li>`).join('')}
-          </ul>
-        </div>
-        
-        <a href="${summary.url}" class="filing-link" target="_blank">View on SEC Website</a>
-      </div>
-    `;
-  });
-  
-  // Add errors if any
-  if (errors.length > 0) {
-    const sanitizedErrors = errors.map(err => ({
-      ticker: sanitizeForEmail(err.ticker),
-      error: sanitizeForEmail(err.error || 'Unknown error')
-    }));
-    
-    html += `
-      <div class="errors">
-        <h3>Issues Encountered</h3>
-        <ul>
-          ${sanitizedErrors.map(err => `<li>${err.ticker}: ${err.error}</li>`).join('')}
-        </ul>
-      </div>
-    `;
-  }
-  
-  // Add footer
-  html += `
-      <div class="footer">
-        <p>This email was generated by tldrSEC. The information provided is for informational purposes only and should not be considered financial advice.</p>
-        <p>© ${new Date().getFullYear()} tldrSEC</p>
-      </div>
-    </body>
-    </html>
-  `;
-  
-  return html;
-}
-
-/**
- * Generate a simple summary based on parsed content
- * This is a fallback when AI summarization is not available
- */
-export function generateSimpleSummary(
-  parsedContent: ParsedContent, 
-  formType: FilingType, 
-  ticker: string, 
-  companyName: string
-): string {
-  const formMetadata = getFormMetadata(formType);
-  const formName = formMetadata ? formMetadata.displayName : formType;
-  
-  let summary = `This is a ${formName} filing from ${companyName} (${ticker}). `;
-  
-  if (parsedContent.title) {
-    summary += `The filing is titled "${parsedContent.title}". `;
-  }
-  
-  if (parsedContent.filingDate) {
-    const filingDate = new Date(parsedContent.filingDate).toLocaleDateString();
-    summary += `It was filed on ${filingDate}. `;
-  }
-  
-  if (parsedContent.sections && parsedContent.sections.length > 0) {
-    summary += `The document contains ${parsedContent.sections.length} major sections. `;
-    
-    // Add information about important sections if available
-    const importantSections = ['Management Discussion', 'Risk Factors', 'Financial Statements'];
-    const foundSections = parsedContent.sections
-      .filter(section => importantSections.some(keyword => section.title.includes(keyword)))
-      .map(section => section.title);
-    
-    if (foundSections.length > 0) {
-      summary += `Key sections include: ${foundSections.join(', ')}. `;
-    }
-  }
-  
-  summary += `Please review the original filing for complete details.`;
-  
-  return summary;
-}
-
-/**
- * Extract key points from parsed content
- */
-export function extractKeyPoints(parsedContent: ParsedContent, formType: FilingType): string[] {
-  const keyPoints: string[] = [];
-  
-  // Try to extract key points from the parsed content
-  if (parsedContent.sections) {
-    // Look for sections that might contain key information
-    const importantSections = parsedContent.sections.filter(section => {
-      const title = section.title.toLowerCase();
-      return (
-        title.includes('summary') ||
-        title.includes('highlight') ||
-        title.includes('overview') ||
-        title.includes('management') ||
-        title.includes('discussion') ||
-        title.includes('analysis') ||
-        title.includes('risk factor') ||
-        title.includes('financial')
-      );
-    });
-    
-    // Extract first paragraph from each important section
-    importantSections.forEach(section => {
-      if (section.content && section.content.length > 0) {
-        // Get first paragraph that's not too short
-        const paragraphs = section.content.split('\n\n');
-        const significantParagraphs = paragraphs.filter(p => p.length > 100);
-        
-        if (significantParagraphs.length > 0) {
-          // Extract a key point from the first significant paragraph
-          const firstPara = significantParagraphs[0];
-          const sentences = firstPara.split(/[.!?]+/).filter(s => s.trim().length > 0);
-          
-          if (sentences.length > 0) {
-            const keyPoint = `${section.title}: ${sentences[0].trim()}.`;
-            keyPoints.push(keyPoint);
-          }
-        }
-      }
-    });
-  }
-  
-  // If we couldn't extract meaningful key points, provide generic ones
-  if (keyPoints.length === 0) {
-    const formMetadata = getFormMetadata(formType);
-    
-    if (formType === '10-K' || formType === '10-K/A') {
-      keyPoints.push('Annual report containing audited financial statements.');
-      keyPoints.push('Provides a comprehensive overview of the company\'s business and financial condition.');
-      keyPoints.push('Includes management\'s discussion and analysis of financial condition and results of operations.');
-    } else if (formType === '10-Q' || formType === '10-Q/A') {
-      keyPoints.push('Quarterly report containing unaudited financial statements.');
-      keyPoints.push('Provides a continuing view of the company\'s financial position during the year.');
-      keyPoints.push('Less detailed than annual reports on Form 10-K.');
-    } else if (formType === '8-K') {
-      keyPoints.push('Current report announcing major events that shareholders should know about.');
-      keyPoints.push('Filed to disclose material events or corporate changes.');
-      keyPoints.push('Review the full filing for specific details about the reported event.');
-    } else {
-      keyPoints.push(`This is a ${formMetadata?.displayName || formType} filing.`);
-      keyPoints.push('Review the full filing for complete details.');
-      keyPoints.push('Consult with a financial advisor for interpretation of this document.');
-    }
-  }
-  
-  return keyPoints;
-}
